% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_threshold.R
\name{absolute_threshold}
\alias{absolute_threshold}
\title{Function to apply a given absolute threshold to the \code{.ncol} input graph}
\usage{
absolute_threshold(
  infile,
  outfile,
  threshold,
  overwrite = FALSE,
  sort_output = FALSE
)
}
\arguments{
\item{infile}{The \code{.ncol} graph file to be thresholded.}

\item{outfile}{The output file path for the thresholded graph to be written to.}

\item{threshold}{The desired absolute threshold to apply to the input graph.}

\item{overwrite}{A boolean parameter meant to prevent overwriting existing thresholded
graph files. The default (\code{FALSE}) will display a menu to the user if the passed
\code{outfile} already exists. Choosing \code{TRUE} will bypass this menu and overwrite
the existing file without interruption from a workflow.}

\item{sort_output}{Sorts the edges of the thresholded graph in descending order before
writing them to \code{outfile}.}
}
\value{
Nothing. The thresholded graph is written to the file specified by outfile.
}
\description{
\code{absolute_threshold()} performs absolute thresholding on an input graph and 
writes the thresholded graph to an output file. This function is tailored specifically
for a clear absolute thresholding use case as compared to \code{threshold()}, which is the
more general purpose thresholding function.
}
\examples{
infile <- system.file('extdata', 'HumanCellCycleSubset.ncol', package = "thresholding") 
thresholding::absolute_threshold(infile, 
                                 outfile = "./HCCS-ABSTHRESH.ncol",
                                 threshold = 0.90,
                                 sort_output = TRUE
                                 )
}
