% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{analysis}
\alias{analysis}
\title{Main graph thresholding analysis function}
\usage{
analysis(
  infile,
  outfile_prefix,
  methods = "",
  lower = 0.5,
  upper = 0.99,
  increment = 0.01,
  window_size = 5L,
  min_partition_size = 10L,
  min_clique_size = 5L,
  min_alpha = 0,
  max_alpha = 4,
  alpha_increment = 0.1,
  num_samples = 0L,
  significance_alpha = 0.01,
  bonferroni_corrected = 0L
)
}
\arguments{
\item{infile}{Name of .ncol graph file to read in for analysis}

\item{outfile_prefix}{Prefix of output file in which analysis will be redirected to (Ex: <PREFIX>.iterative.txt )}

\item{methods}{Comma separated list of analysis methods, listed if thresholding::help() is called (defaults to none)}

\item{lower}{lower bound to begin thresholding increment (lower >= 0)}

\item{upper}{hard upper bound that ends thresholding loop}

\item{increment}{size of step in the thresholding loop}

\item{window_size}{DOCUMENT THIS}

\item{min_partition_size}{DOCUMENT THIS}

\item{min_clique_size}{DOCUMENT THIS}

\item{min_alpha}{DOCUMENT THIS}

\item{max_alpha}{DOCUMENT THIS}

\item{alpha_increment}{DOCUMENT THIS}

\item{num_samples}{DOCUMENT THIS}

\item{significance_alpha}{DOCUMENT THIS}

\item{bonferroni_corrected}{DOCUMENT THIS}
}
\description{
Main graph thresholding analysis function
}
