% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{edge_hist}
\alias{edge_hist}
\title{edge_hist}
\usage{
edge_hist(infile, bin_width, sep = "\\t")
}
\arguments{
\item{infile}{tab separated .ncol formatted graph file (list of weighted edges - wel)}

\item{bin_width}{width of each histogram bin}

\item{sep}{determines how the wel file is separated (default is tab separated file)}
}
\description{
Histogram function for displaying edge weight frequencies
}
