% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_iter_t_vals}
\alias{get_iter_t_vals}
\title{get_iter_t_vals()
User wrapper function for get_iterative_t_values
Returns the thresholding data frame created by the internal
get_iterative_t_values, which does not return anything to the user.}
\usage{
get_iter_t_vals(outfile_prefix)
}
\arguments{
\item{outfile_prefix}{Prefix of output file, which can have several
output file paths if the same prefix is run several. This assumes
that the desired file(s) are in the current working directory
(pwd)

Ex.) get_iter_t_vals("iter-prefix") will execute get_iterative_t_values
on all files names iter-prefix-###.iterative.txt where ### is the process
ID of the internal process (like 4318, 3341, 414143, etc.).}
}
\description{
The returned data frame includes graph and graph method values for each 
increment of the threshold.
}
