% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_iterative_t_values}
\alias{get_iterative_t_values}
\title{Translated from Carissa Bleker's Thresholding helper functions
for quicker analysis of thresholding analysis results}
\usage{
get_iterative_t_values(files, D = NULL, d_min_t = list(general = 0))
}
\arguments{
\item{files}{Vector of file paths or just one file path}

\item{D}{Optional argument that accepts a list. This can be used
to capture the thresholding results by a certain method. If just
using get_iter_t_vals, then D can be left unspecified. Otherwise,
get_results fills out and returns this value automatically}

\item{d_min_t}{INTERNAL CONTROL}
}
\description{
The link to Carissa's Github repo: 
  https://github.com/carissableker/thresholding
The link to the combine_analysis_results functions: 
  https://github.com/carissableker/thresholding/blob/master/example/combine_analysis_results.py
Helper function for thresholding::get_results()
Gets thresholding values from the iterative results
Reads in tab the separated .iterative.txt files from 
thresholding::analysis
}
