% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-results-SINGLEFILE.R, R/get_results.R
\name{get_results}
\alias{get_results}
\title{Prints the resulting analysis method thresholding values
after running thresholding::analysis()}
\usage{
get_results(outfile_prefix, plot_iterative = FALSE)

get_results(outfile_prefix, plot_iterative = FALSE)
}
\arguments{
\item{outfile_prefix}{filename or file path for resulting output file from 
running the analysis function (file would be <prefix>.iterative.txt)}

\item{plot_iterative}{optionally plot the vertices and edges vs. threshold value as a graph.
uses ggplot2 to automatically call this package's plot_t_vs_ev() function without
the user needing to manually extract the required parameters.}

\item{outfile}{filename or file path for resulting output file from 
running the analysis function (file would be <prefix>.<type>.txt where the 
prefix is the name of the output file before the extension and the 
type could be "iterative", "statistical_errors", or "local_global".
Example: dataset-567.iterative.txt
     prefix: datset-567
     extension: .iterative.txt}
}
\description{
Note: This function assumes that the .iterative anaylsis() output file
is in the current directory unless the path to the file is provided.

Note: This function assumes that the .iterative anaylsis() output file
is in the current directory unless the path to the file is provided.
}
