% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_sig_t_vals}
\alias{get_sig_t_vals}
\title{get_sig_t_vals()
User wrapper function for \code{get_significance_t_values()}
Returns the thresholding data frame created by the internal
\code{get_significance_t_values()}, which does not return anything to the user.}
\usage{
get_sig_t_vals(outfile_prefix, recursive = FALSE)
}
\arguments{
\item{outfile_prefix}{Prefix of output file, which can have several
output file paths if the same prefix is run with analysis() several times. 

Ex.) get_iter_t_vals("iter-prefix") will execute get_iterative_t_values
on all files names iter-prefix-###.iterative.txt where ### is the process
ID of the internal process (like 4318, 3341, 414143, etc.).}

\item{recursive}{Option to allow for subdirectory searching of the file prefix.
Set to FALSE for default in case of large subdirectories. This option is meant to help 
save time with R Studio and other editors' auto completion feature, which allows for tabbing
to complete the file name from the present working directory (PWD) and maintaining the current
PWD. 

If \code{FALSE}, \code{get_sig_t_vals()} will use the path passed to the function from outfile_prefix.
This may be either relative or absolute. Only the directory containing the desired file(s)
will be searched. (Note: if no file paths are explicitly provided, the PWD (".") will be used.)

If \code{TRUE}, \code{get_sig_t_vals()} will recursively searching through all files matching the Regex pattern
of \code{^output_prefix.*\\.iterative\\.txt$}, which essentially allows for any file name that starts 
with the specified outfile_prefix and ends with ".iterative.txt".}
}
\description{
The returned data frame includes graph and graph method values for each 
increment of the threshold.
}
