% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_t_vs_ev.R
\name{plot_t_vs_ev}
\alias{plot_t_vs_ev}
\title{Plots the edge (E) and vertex (V) counts of a graph at varying thresholds}
\usage{
plot_t_vs_ev(iter_prefix)
}
\arguments{
\item{iter_prefix}{The prefix of the .iterative.txt file(s) from the \code{analysis()} function.}
}
\value{
A ggplot object that can be displayed with show(), or by calling the plot_t_vs_ev() directly in the R terminal.
}
\description{
Plots a graph's edge and vertices by threshold value on two different scales.
Additionally, non-NaN and non-Inf analysis methods are plotted at their optimal thresholds \
against the V/E line plots.
}
