% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{thresh.analysis}
\alias{thresh.analysis}
\title{Thresholding Analysis
Thresholding functions - documentation handled through roxygen2 style comments}
\usage{
thresh.analysis(
  outfile_prefix,
  G,
  l = 0.5,
  u = 0.99,
  increment = 0.01,
  windowSize = 5,
  minimumpartitionsize = 10,
  minimum_cliquesize = 5,
  min_alpha = 0,
  max_alpha = 4,
  alpha_increment = 0.1,
  num_samples = 0,
  significance_alpha = 0.01,
  bonferroni_corrected = 0,
  methods
)
}
\arguments{
\item{outfile_prefix}{Prefix to outfile (Ex. "test." turns to "test.iterative.txt")}
}
\description{
thresh.analysis - Primary Thresholding Analysis Function

TODO: ADD PARAM DOCUMENTATION / REMOVE FROM FUNCTION PARAMS
}
