% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{ThresholdTuner-package}
\alias{ThresholdTuner}
\alias{ThresholdTuner-package}
\title{ThresholdTuner: Tools for Graph Analysis, Threshold Selection, and Thresholding}
\description{
An R wrapper for C++ code that provides multiple graph thresholding utilities and analysis tools. It implements a series of algorithms to identify a statistically significant threshold for edge weights, which can then be applied to remove edges below the threshold, simplifying the graph and pruning disconnected vertices. The methods are applicable to biological data, correlation networks (e.g., Pearson and Spearman), and other graph types. The tools accept a graph input file, and the user specifies the methods to determine an optimal threshold. The thresholding tools rely on the 'igraph' C library for graph construction and manipulation, and the 'alglib' library for mathematical computations. Upon completing the analysis, users can visualize results, extract thresholds, and export a thresholded graph to a separate graph file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/snapplelevi/langston-thresholding}
  \item \url{https://github.com/carissableker/thresholding}
  \item Report bugs at \url{https://github.com/snapplelevi/langston-thresholding/issues}
}

}
\author{
\strong{Maintainer}: Levi Hochstetler \email{levijhoch@gmail.com}

Authors:
\itemize{
  \item Carissa Bleker
  \item Stephen K. Grady
  \item Michael A. Langston
}

}
\keyword{internal}
