% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_threshold.R
\name{absolute_threshold}
\alias{absolute_threshold}
\title{Function to apply a given absolute threshold to the \code{.ncol} input graph}
\usage{
absolute_threshold(
  infile,
  outfile,
  threshold,
  overwrite = FALSE,
  sort_output = FALSE
)
}
\arguments{
\item{infile}{string. The path to the \code{.ncol} graph file to be thresholded.}

\item{outfile}{string. The output file path for the thresholded graph to be written to.}

\item{threshold}{numeric. The desired absolute threshold to apply to the input graph.}

\item{overwrite}{boolean. A parameter meant to prevent overwriting existing thresholded
graph files. The default (\code{FALSE}) will display a menu to the user if the passed
\code{outfile} already exists. Choosing \code{TRUE} will bypass this menu and overwrite
the existing file without interruption from a workflow.}

\item{sort_output}{boolean. Sorts the edges of the thresholded graph in descending order before
writing them to \code{outfile} if set to \code{TRUE}. Otherwise, descending order is not guaranteed.}
}
\value{
Nothing. The thresholded graph is written to the file specified by outfile.
}
\description{
\code{absolute_threshold()} performs absolute thresholding on an input graph and 
writes the thresholded graph to an output file. This function is tailored specifically
for a clear absolute thresholding use case as compared to \code{threshold()}, which is the
more general purpose thresholding function.
}
\examples{
library(thresholding)
infile <- system.file('extdata', 'HumanCellCycleSubset.ncol', package = "thresholding") 
outfile <- tempfile(fileext = ".ncol")
absolute_threshold(infile, 
                  outfile = outfile,
                  threshold = 0.90,
                  sort_output = TRUE,
                  overwrite = TRUE
                  )
print(paste0("Thresholded graph in '", outfile, ".'"))
}
