% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{analysis}
\alias{analysis}
\title{Analysis methods for absolute thresholding of weighted graphs}
\usage{
analysis(
  infile,
  methods = as.numeric(c()),
  outfile_prefix = "",
  lower = 0.5,
  upper = 0.99,
  increment = 0.01,
  window_size = 5L,
  min_partition_size = 10L,
  min_clique_size = 5L,
  min_alpha = 0,
  max_alpha = 4,
  alpha_increment = 0.1,
  num_samples = 0L,
  significance_alpha = 0.01,
  bonferroni_corrected = 0L,
  overwrite = FALSE
)
}
\arguments{
\item{infile}{File path for .ncol graph file (\link{https://lgl.sourceforge.net/}) to read in for analysis. 
This file must be space delimited for this function to properly read in the graph's information.}

\item{methods}{Numeric vector of method integers. Defaults to an empty list. The number to method translation is given below:
\itemize{
  \item 0 = all
  \item 1 = significance and power calculations (only valid for Pearson CC)
  \item 2 = local-global
  \item 3 = scale free
  \item 4 = maximal cliques
  \item 5 = spectral methods
  \item 6 = random matrix theory
  \item 7 = clustering coefficient
  \item 8 = percolation
}
 Refer to the following papers for more detailed description of these methods: 
\itemize{
\item Carissa Bleker's thresholding dissertation: \link{https://trace.tennessee.edu/utk_graddiss/5894/} 
\item Dr. Langston, Grady, and Bleker's thresholding paper: \link{https://web.eecs.utk.edu/~mlangsto/JCB-Thresholding-Paper.pdf}
}}

\item{outfile_prefix}{Prefix of output file in which analysis will be redirected to. If this is not specified,
\code{thresholding::analysis()} will auto generate the output file prefix to include the input file's 
prefix and the ascending method numbers. 
The input file prefix will be determined by the characters preceding the first period ('.') character.
  An example if the user requests methods \code{4} and \code{7} with an input file named \code{"myfile.tsv"} would be:
     \code{myfile-47.<method_name>.txt}
 Method name can vary based on the methods used. This will either be \code{iterative}, \code{local_global}, or \code{statistical_errors}.}

\item{lower}{Lower bound to begin thresholding loop at (default = 0.5 ; lower >= 0)}

\item{upper}{Hard upper bound that ends thresholding  loop when \code{lower} value is greater than \code{upper} value (Default = 0.99)}

\item{increment}{Size of increment step in the thresholding loop}

\item{window_size}{Sliding window size for spectral method (Default = 5)}

\item{min_partition_size}{minimum size of graph or subgraph after threshold (Default = 10)}

\item{min_clique_size}{Minimum size of maximal cliques in maximal clique count (Default = 5)}

\item{min_alpha}{Starting alpha value used in \strong{method 2 - local global pruning}.  \emph{Not used in any other methods.}  (Default = 0.0)}

\item{max_alpha}{Ending alpha value used in \strong{method 2 - local global pruning}.  \emph{Not used in any other methods.}   (Default = 4.0)}

\item{alpha_increment}{Size of increment of alpha value in \strong{method 2 - local global pruning}'s main loop. \emph{Not used in any other methods.}  (Default = 0.1)}

\item{num_samples}{Number of samples in Pearson Correlation Coefficient data (only used for \strong{analysis method 1 - Power and Significance calculations}).
\emph{\strong{\code{num_samples} must be positive, non-zero, and match the number of samples from the original dataset for method 1 to work.}} \emph{Not used in any other methods.}}

\item{significance_alpha}{Probability of rejecting the null hypothesis when the null hypothesis is true.  (Default = 0.01)}

\item{bonferroni_corrected}{Option to perform Bonferroni correction in \strong{method 1 - significance and power calculations}. 
Applies Bonferroni correction to the value of significance_alpha if set to \code{TRUE}. \emph{Not used in any other methods.} (default \code{FALSE})}

\item{overwrite}{Determines whether output file with given or generated prefix will be overwritten. 
Set this to \code{TRUE} to force overwrite the output file. The default, \code{FALSE}, will display a menu asking
whether or not you wish to overwrite the output file.}
}
\value{
Nothing. \code{analysis()} writes all output to a file. The file path and file prefixes are printed on standard output when `analysis()` terminates.
}
\description{
Weighted graphs thresholding analysis. This function iteratively performs absolute thresholding the input graph at specified threshold values [\code{lower}, \code{upper}].
The function's methods parameter controls which graph analysis methods are performed at each threshold value.
Execution of \code{analysis()} ends after the '\code{upper}' threshold is reached, or if the graph becomes
too small to threshold further.
 
The results at each step are written output files. There is always at least one output file, but there can be 
up to three depending on the methods passed:
\itemize{
  \item \code{<outfile_prefix>.iterative.txt}:          (\strong{guaranteed})
  \item \code{<outfile_prefix>.statistical_errors.txt}: (method \strong{1})
  \item \code{<outfile_prefix>.local_global.txt}:        (method \strong{2})
}
}
\details{
Refer to Dr. Carissa Bleker's dissertation for more information about these analysis methods: \link{https://trace.tennessee.edu/utk_graddiss/5894/}
}
\examples{
#######    Variable Set-Up     #######
data_file <- system.file('extdata', 'HumanCellCycleSubset.ncol', package = "thresholding")   # .ncol weighted edge list
data_prefix <- './example/HumanCellCycleSubset-thresh'  # prefix used for output file(s)
lower <- 0.6   

#######    Example 1 - No methods #######
analysis(data_file, 
         data_prefix,
         lower = lower,
         )

#######    Example 2 - Iterative methods #######
# WARNING FOR EXAMPLES: this code may take a few minutes to run.
data_file <- system.file('extdata', 'HumanCellCycleSubset.ncol', package = "thresholding") 
methods <- c(3,4,5,6)
lower <- 0.7

# Note: By supplying an output file prefix, the only modification to the output
#       file will be "-###" where ### represents the analysis methods passed
#       to analysis(). This is done to distinguish between output files generated by analysis()
#       that have the same prefix, but potentially different methods.
analysis(data_file,
         outfile_prefix = "./example_2_it",
         methods = methods,
         lower = lower,
        )


#######    Example 3 - iterative and power/significance methods #######
data_file <- system.file('extdata', 'HumanCellCycleSubset.ncol', package = "thresholding") 
methods <- c(8, 1, 3)    # select the three desired analysis methods
lower <- 0.6             # choose lower bound thresholding value the thresholding loop begins at
num_samples <- 13        # ONLY FOR METHOD 1 - number of samples in data set

# Note: analysis() will autogenerate an output file name based 
#       on the input file path if a prefix is not passed.
analysis(data_file, 
         methods = methods,
         lower = lower,
         num_samples = num_samples,
         )
}
