% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{analysis}
\alias{analysis}
\title{Analysis methods for absolute thresholding of weighted graphs}
\usage{
analysis(
  infile,
  methods = as.numeric(c()),
  outfile_prefix = "",
  lower = 0.5,
  upper = 0.99,
  increment = 0.01,
  window_size = 5L,
  min_partition_size = 10L,
  min_clique_size = 5L,
  min_alpha = 0,
  max_alpha = 4,
  alpha_increment = 0.1,
  num_samples = 0L,
  significance_alpha = 0.01,
  bonferroni_corrected = 0L,
  overwrite = FALSE
)
}
\arguments{
\item{infile}{File path for .ncol (\link{https://lgl.sourceforge.net/}) graph file to read in for analysis. 
This file must be space delimited for this function to properly read in the graph's information.}

\item{methods}{Numeric vector of method integers. Defaults to an empty list. The number to method translation is given below:
\itemize{
  \item 0 = all
  \item 1 = significance and power calculations (only valid for Pearson CC)
  \item 2 = local-global
  \item 3 = scale free
  \item 4 = maximal cliques
  \item 5 = spectral methods
  \item 6 = random matrix theory
  \item 7 = clustering coefficient
  \item 8 = percolation
}
 Refer to the following papers for more detailed description of these methods: 
\itemize{
\item Carissa Bleker's thresholding dissertation: \link{https://trace.tennessee.edu/utk_graddiss/5894/} 
\item Dr. Langston, Grady, and Bleker's thresholding paper: \link{https://web.eecs.utk.edu/~mlangsto/JCB-Thresholding-Paper.pdf}
}}

\item{outfile_prefix}{Prefix of output file in which analysis will be redirected to. If this is not specified,
thresholding::analysis() will auto generate the output file prefix to include the input file's 
prefix and the ascending method numbers. 
The input file prefix will be determined by the characters preceding the first period ('.') character.
  An example if the user requests methods \code{4} and \code{7} with an input file named \code{"myfile.tsv"} would be:
     \code{myfile-47.<method_name>.txt}
 Method name can vary based on the methods used. This will either be iterative, local_global, or statistical_errors.}

\item{lower}{Lower bound to begin thresholding loop at (default 0.5 ; lower >= 0)}

\item{upper}{Hard upper bound that ends thresholding  loop when `lower` value is greater than `upper` value (default 0.99)}

\item{increment}{Size of increment step in the thresholding loop}

\item{window_size}{Sliding window size for spectral method (default is 5)}

\item{min_partition_size}{minimum size of graph or subgraph after threshold (default is 10)}

\item{min_clique_size}{minimum size of maximal cliques in maximal clique count (defaultiis 5)}

\item{min_alpha}{DOCUMENT THIS}

\item{max_alpha}{DOCUMENT THIS}

\item{alpha_increment}{DOCUMENT THIS}

\item{num_samples}{number of samples in Pearson Correlation Coefficient data (only used for analysis method 1 - Power and significance calculations)}

\item{significance_alpha}{DOCUMENT THIS}

\item{bonferroni_corrected}{switch to perform bonferroni corrections in significance and power calculations (default FALSE)}

\item{overwrite}{Determines whether output file with given or generated prefix will be overwritten. 
Set this to 'TRUE' to force overwrite the output file. The default, 'FALSE', will display a menu asking
whether or not you wish to overwrite the output file.}
}
\value{
Nothing. `analysis()` writes all output to a file. The file path and file prefixes are printed on standard output when `analysis()` terminates.
}
\description{
Weighted graphs thresholding analysis. This function iteratively performs absolute thresholding the input graph at specified threshold values [lower, upper].
The function's methods parameter controls which graph analysis methods are performed at each threshold value.
Execution of analysis() ends after the 'upper' threshold is reached, or if the graph becomes
too small to threshold further.
 
The results at each step are written output files. There is always at least one output file, but there can be 
up to three depending on the methods passed:
\itemize{
  \item <outfile_prefix>.iterative.txt:          (\strong{guaranteed})
  \item <outfile_prefix>.statistical_errors.txt: (method \strong{1})
  \item <outfile_prefix>.local_global.txt        (method \strong{2})
}
}
\details{
Refer to Dr. Carissa Bleker's dissertation for more information about these analysis methods: \link{https://trace.tennessee.edu/utk_graddiss/5894/}
}
\examples{
#######    Variable Set-Up     #######
data_file <- './example/HumanCellCycleSubset.ncol'      # .ncol weighted edge list
data_prefix <- './example/HumanCellCycleSubset-thresh'  # prefix used for output file(s)
lower <- 0.6   

#######    Example 1 - No methods #######
analysis(data_file, 
         data_prefix,
         lower = lower,
         )

#######    Example 2 - Iterative methods #######
print("COMING SOON 2")

#######    Example 3 - iterative and power/significance methods #######
methods <- c(8, 1, 3)    # select the three desired analysis methods
lower <- 0.6             # choose lower bound thresholding value the thresholding loop begins at
num_samples <- 13        # ONLY FOR METHOD 1 - number of samples in data set

# Note: analysis() will autogenerate an output file name based 
#       on the input file path if a prefix is not passed.
analysis(data_file, 
         methods = methods,
         lower = lower,
         num_samples = num_samples,
         )

######    Example 4 - iterative and local-global  methods #######
print("COMING SOON 4")
}
