% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results}
\alias{get_results}
\title{Prints the resulting analysis method thresholding values
after running thresholding::analysis()}
\usage{
get_results(outfile_prefix, plot_iterative = FALSE)
}
\arguments{
\item{outfile_prefix}{filename or file path for resulting output file from 
running the analysis function (file would be <prefix>.iterative.txt)}
}
\description{
Note: This function assumes that the .iterative anaylsis() output file
is in the current directory unless the path to the file is provided.
}
