% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_t_vs_ev.R
\name{plot_t_vs_ev}
\alias{plot_t_vs_ev}
\title{Plots the edge (E) and vertex (V) counts of a graph at varying thresholds}
\usage{
plot_t_vs_ev(iter_prefix)
}
\arguments{
\item{iter_prefix}{string. The prefix of the .iterative.txt file(s) from the output \code{analysis()} function, which
is printed to the screen after \code{analysis()} ends.}
}
\value{
A ggplot object that can be displayed with show() (can be stored or plotted directly by calling  \code{show(plot_t_vs_ev())} directly in the R terminal.)
}
\description{
Plots a graph's edge and vertices by threshold value on two different scales. This function is intended to 
give more visual insight to the thresholds chosen by the iterative analysis methods of \code{analysis()},
which are methods 3 through 8.
}
\details{
Additionally, non-NaN and non-Inf analysis methods are plotted at their optimal thresholds
against the V/E line plots.


The vertical, orange lines represent a thresholding analysis method. The intersection point of both the V and E
thresholding curves are denoted by red and blue squares. The method name is annotated for each vertical line. The x value (s)
(threshold amount) that the method has yielded are where the vertical lines will appear. Use \code{get_results()} to see the 
concrete value of these thresholds.
}
