% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_t_vs_ev.R
\name{plot_t_vs_ev}
\alias{plot_t_vs_ev}
\title{Plots the edge (E) and vertex (V)count of a graph at varying thresholds
Marks non-null methods and their optimal thresholds against the 
V/E line plots}
\usage{
plot_t_vs_ev(iter_infile)
}
\arguments{
\item{iter_infile}{The name of the .iterative.txt file from the \code{analysis()} function.
This is required to be the exact name of the file because \code{plot_t_vs_ev()} will only show
the methods used for that specific combination of methods. In comparison, \code{get_results()} 
takes in a prefix so that it can potentially combine analysis methods for several analysis
method combinations that share the same user given prefix.}
}
\description{
Plots the edge (E) and vertex (V)count of a graph at varying thresholds
Marks non-null methods and their optimal thresholds against the 
V/E line plots
}
