% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{threshold}
\alias{threshold}
\title{Strict thresholding for weighted graphs in .ncol format}
\usage{
threshold(
  infile,
  outfile,
  method = "absolute",
  absolute = 0,
  local_global_alpha = 0,
  rank = 0L
)
}
\arguments{
\item{infile}{DOCUMENT THIS}

\item{outfile}{DOCUMENT THIS}

\item{method}{DOCUMENT / CHANGE THIS}

\item{absolute}{DOCUMENT / CHANGE THIS}

\item{local_global_alpha}{DOCUMENT / CHANGE THIS}

\item{rank}{DOCUMENT / CHANGE THIS}
}
\value{
Nothing. The thresholded graph is written to the file specified by outfile.
}
\description{
Description of threshold goes here
}
\examples{
print('NEED TO WRITE EXAMPLES FOR THIS LATER')
print('MULTIPLE IF POSSIBLE')
}
